tests: add certificate revocation list tests.
This commit adds tests that ensure rcgen created CRLs parse/validate
successfully with OpenSSL and x509-parser.

Coverage is not added for Botan because it appears the Botan rust
bindings don't expose the required `X509_CRL` type.

Coverage is not (yet) added for Webpki because the upstream crate hasn't
yet cut a release that includes CRL parsing support.