ext: introduce module for X.509 extension handling
This commit creates a new crate-internal module, `ext`, for managing
X.509 extensions. In this commit we wire up emitting extensions managed
by this module, but do not yet convert any existing extensions to the
new arrangement. This will begin in subsequent commits.

Notably the new representation for a collection of extensions uses
a `BTreeMap` keyed by OID. We use this to reject additions of extensions
that are already present in the map, since RFC 5280 forbids duplicate
extensions. Using a `BTreeMap` instead of another map type allows
preserving order.