ext: implement subject key ID, specifying SKI
This commit lifts the subject key identifier extension into the `ext`
module.

It additionally adds support for specifying a custom SKI value in
`CertificateParams`, and reading it from a CSR's SKI ext.

Diverging from the existing code we now adhere to the RFC 5280 advice
and always emit the SKI extension when generating a certificate.
Previously this was only done if the basic constraints specified
`IsCa::Ca` or `IsCa::ExplicitNoCa`, but not when using `IsCa::NoCa`.