lib: use `Extensions` to decide when to emit exts
Previously the params were interrogated for parameters that would
indicate needing an extension to be emitted. Keeping this logic up to
date was error prone. This commit reworks the code to emit extensions
whenever the `Extensions` generated from the params isn't empty. This
has the advantage of not needing to be updated when new parameter ->
extension instances are added.