examples: fix typo, include ClientAuth EKU
There was a request for an example program that:

1. Generates a self-signed CA
2. Issues a certificate for a client
3. Issues a certificate for the server

The `sign-leaf-with-ca.rs` example is close to meeting this need. It
accomplishes 1 and 3, but because of the EKU the generated EE cert can't
be used for 2.

This commit adds the client auth EKU and now it should be feasible to
use the EE cert for both the client and server, or with minimal tweaks,
to make two separate EE certs one for each purpose.