lib: fix Certificate comment
The `Certificate` type as it exists at this commit has no signature, and
it isn't self-signed. It _may_ be turned into a self signed certificate
by calling `serialize_der()` or `serialize_pem()`, or it may be turned
into a certificate signed by an issuer! Simplify the comment before more
substantial refactoring.