add more flexible fn for borrowing cert DER
The `Certificate::der()` function was written to return
`&CertificateDer<'static>` - that is, a reference to the certificate's
owned DER.

In practice this makes it challenging for downstream users that may want
a `CertificateDer<'static>`, that is, their own owned certificate DER,
not a reference to one.

This commit adds a second `borrow_der()` fn to avoid making a breaking
release to fix this. The existing `der()` fn is updated to encourage
using `borrow_der()`. This gives you a `CertificateDer<'_>` which is
equivalent for all the current use cases, but also supports calling
`CertificateDer::into_owned()` to get a `CertificateDer<'static>` of
your own when required.

Unit tests are updated throughout to prefer `borrow_der()`.